Added support for scenario-based flavor creation.
As certain scenarios such as, DPDK, FDIO, and OVS use cases
require special 'hw:' settings to flavors, SNAPS-OO has just
implemented the ability to take in these configuration values
to apply to its integration tests. This change checks the
scenario and should be configuring the flavor metadata accordingly.

Also see SNAPS-38

Used this as an opportunity to refactor the SNAPS test classes as well...

JIRA: FUNCTEST-746

Change-Id: I63535776acde088a7461f602966a324b6207592d
Signed-off-by: spisarski <s.pisarski@cablelabs.com>