Update patch set 4
Patch Set 4:

> > Switching to UEFI implies, besides adding the new properties,
 > > removing kernel and ramdisk from config, so this change is not
 > > needed.
 > > If someone wants to stick to 3-part, they should add back
 > > krn/ramdisk, but there is no need for us to keep them if UEFI
 > works
 > > well.
 > 
 > I was under the impression we will make the 1-part /3-part image
 > configurable. If that's off the table all we need is to remove
 > kernel and ramdisk from the functest config yaml

I was thinking this is configurable enough - if the user wants 3-part back, they should remove hw_firmware_type/short_id and add the kernel/initrd (with this patch they would only have to remove the 2 properties to swtich to 3-part).
Since we are trying to move everything to UEFI on AArch64, keeping the kernel/initrd URLs would only be confusing.

Patch-set: 4