Update patch set 1
Patch Set 1:

> (1 comment)

 > (1 comment)

it is not the only change to be done
as I would like to see the feature supported in Danube, I just considered Helen's comment. 

I agree to change that after Danube for get_db_name and all the other get_(pod/installer/..) => use CONST.

but the functions are still used in several files (e.g. onos)
this change could be relatively big (several files + unit tests) and risky regarding the timeline, I prefer to create a JIRA and to do it out of Dnaube rush

Patch-set: 1
Label: Verified=0