Update patch set 1
Patch Set 1: Code-Review+1

This patch is serving a similar purpose as https://gerrit.opnfv.org/gerrit/#/c/29701/. Setting the selected modules as WARNING is a quick way to fix the logging chaos.
Ideally, the selected modules will be one level higher than the Functest logger could be a complete solution which is indicated in patch 29701.

Patch-set: 1
Label: Code-Review=+1