Update patch set 17
Patch Set 17:

(3 comments)

I understand we are at a critical point of release.
I agree with putting it into a later phase, such as Danube 2.0 or later.
Also, I am not sure the impact to the opera release.

The shared functions(clearwater rest apis) of cloudify and opera is fully tested as I set up a cloudify vIMS locally.

Patch-set: 17