Revert "Make SFC test a python call to main()"
This reverts commit d5820bef80ea4bdb871380dbfe41db12290fc5f8.

Robot test runs before SFC test and it imports
https://github.com/robotframework/SSHLibrary
which does a monkey patching in 
the python runtime / paramiko.
Untill now sfc run in a new python process (clean)
because it run using the bash command.
But when importing it as a module and call main()
from python, it will run in the patched runtime 
and it will error out.
https://hastebin.com/iyobuxutib.py


Change-Id: I54237c32c957718b363d302efe84e01bc78e4f47
Signed-off-by: George Paraskevopoulos <geopar@intracom-telecom.com>