Bugfixing the testcases execusion
tempest.api.compute.servers.test_server_actions\
.ServerActionsTestJSON.test_reboot_server_hard
is working in fuel so do test it.

The "-" was missing so
tempest.scenario.test_server_basic_ops
.TestServerBasicOps.test_server_basic_ops
was executed on fuel which is known to be
not working since metadata service is not
available. That was added already long
ago but through the not adding
of the "-" it got added again.

Change-Id: I6808ce9283271e2d2f8acf7c2587aed006d1cbc2
Signed-off-by: Nikolas Hermanns <nikolas.hermanns@ericsson.com>
(cherry picked from commit f85b54b8fc8843897f9c1a7b4711b489ab29f297)