Manage criteria in TestCase
It converts all criteria values to the corresponding percent in
functest/ci/testcases.yaml. It fixes tier_handler.py which required
that type criteria was str.

It is compatible with the old behavior but warns to update.
It will allow a safer remove.

It also fixes a bug in test_tempest to allow merging [1].

[1] https://gerrit.opnfv.org/gerrit/#/c/27949/

Change-Id: Ib6edcfa3103b7d51b0bdc83119f1cea2a8be9fbc
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>