Add a new directory to hold pre-downloaded images
All images required by Functest other than cirros are put under
${HOME}/functest/data/custom_images
In this way, if users want to run Functest offline, they can
pre-download the required images to local and map the local folder to
the directory.

The original cirros images are kept as-is under
${HOME}/functest/data as they are downloaded during docker image
building. If putting them in the new directory, the images will be lost
during docker volume mapping.

Except for the cirros images, any other images need to be placed in
the new directory.

Change-Id: I0d309b93f52b3da23d6130056f1a19907313ef68
Signed-off-by: helenyao <yaohelan@huawei.com>