Update patch set 2
Patch Set 2: Code-Review-1

I don't think this is the way to go for offline testing. My idea was to attach the needed pre-downloaded images as a docker volume if you want to run functest offline. So, it should be documented where to download the images from and how to attach them in the container (maybe an automated script that does all that for you).

Patch-set: 2
Label: Code-Review=-1