Update patch set 4
Patch Set 4:

It works as we agreed in Paris last week. 
 - only core and ci will print info (and above) messages in console.
 - all warn or error messages are printed in console
 - all debug messages are printed to functest.log
Please see WARNING - Please update result which must be an int!.

I strongly believe that CI_DEBUG is useless and it's better to switch any LEVEL in logging.ini. You are free to change any level of packages in your dev env. The logging.ini proposed is much better from a CI point of view.

The key change is not logging.ini which is quite trivial comparing to remove all dependencies to the ft_logger.

Patch-set: 4