Update patch set 2
Patch Set 2:

> but .... it means that we should cherry pick all the new
 > abstraction changes for Danube 2.0...as it is adopting the new
 > design available on master
 > otherwise a patch on the timeout only directly on danube is
 > possible but it will become very messy...
 > related to the discussion we have during the weekly meeting..
 > @Jose I think it will be impossible not to remerge master to
 > Danube...

If this patch put us in dilemma, I would suggest to split this patch. Put the timeout in danube 2.0 as the CI performance is not very good and it needs the timeout to carry on.

Also, the opera daily job should run against Danube instead of latest if we cannot put the new abstraction into Danube 2.0.

Patch-set: 2
Label: Verified=0