Update patch set 4
Patch Set 4:

Only printing the ci & core message in console can make the printout neat and clean, however, it makes the debug very hard for CI issue. 

Before pushing this patch to master, we have to firstly make sure the functest.log will be pushed to artifact no matter the job failed/timeout. 

I would suggest no rush for this patch. Also, I would say, please put the logger_collector before this patch. If the logger_collector is proven to work as expected, we then think about put most of the message for test cases to functest.log.
Still, I think the output(warn/info) of test cases are very useful and need to be print out to console.

Patch-set: 4