Update patch set 2
Patch Set 2:

> the directory /images sounds better to store all the pre-downloaded
 > images. 100% agree with that.
 > 
 > There is no problem in SDNVPN, I am committer there and can do any
 > change to adapt to Functest framework. Other committers will agree
 > with it.

In addition to the new /images folder, a script to download all required images is also needed. Users have to run the script when they have network access and the images will be under /images, then they continue to run offline. 

In this way, it will decrease the chance that users put the images in wrong folder and save their time to read documents.

Patch-set: 2
Label: Verified=0