Update patch set 1
Patch Set 1:

> Please see the similar change
 > https://gerrit.opnfv.org/gerrit/#/c/34479/
 > 
 > We should only get the total of crititical test failed. Could you
 > please check if the return value doesn't count all tests ? And if
 > an higher value (eg 255) is returned when failure.
 > 
 > I prefer a ratio here but we are not obliged and 22 can be set as
 > criteria.

ok, so 34479 is on a better track than this. I'll abandon this for
that.

I was off with 255. we can only get up to 250 as a "good" return
code. If more than 250 tests have passed, we'll still get 250
back. All return codes from 251 - 255 are for specific error 
reasons (missing file, etc).

I didn't follow the point about "not obliged and 22 can be set as
criteria"

Patch-set: 1