Update patch set 1
Patch Set 1:

This is a good principle but I think we can't apply it here.
We will have to apply one instruction in any case (import a module/call a function...).

I'm thinking about disabling directly the logger of the tested class but I'm not sure it currently works in case of test_feature.py which mocks time(). But here it's just a bugfix and I will apply the second step later if possible (very useful only if functest run unit tests as a testcase)

Patch-set: 1