Update patch set 5
Patch Set 5: Code-Review-1

Why are you removing clean_flag from testcases.yaml? Now it's always true. That flag was meant for avoiding taking snapshot+clean for the test cases that it is not needed. This helps saving some time.

Agree with Morgan, having the clean() in TestCase is the right approach, however the goal is to get rid of openstack snapshot + clean by using SNAPS.

Patch-set: 5
Label: Code-Review=-1