Export the Pylint results via pylint.out
It allows integrating Pylint with Jenkins (pylint.out can be defined
as the "XML filename pattern" after installing the Violations plugin
[1]).

It's considered as fine to locally disable warnings such as docstrings
in unit tests. No pylint rcfile is defined for only two options.

Tox still prints the partial results in console to be interpreted by
users. The persistent flag is removed to conform with nosetests args.

[1] https://wiki.jenkins-ci.org/display/JENKINS/Violations

Change-Id: I21662b27c8e4b61e65c4c3f5aab968e3e7dd6d90
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>