Update patch set 3
Patch Set 3:

The test named test_create_three_part_image_from_file_3_creators will need to have the 3 part forms also downloaded and have their file locations configured as well for offline testing. It is also an option to omit some of these tests from Functest if desired. We can discuss as a Functest team which SNAPS tests we feel are not as important.

Patch-set: 3