Update patch set 2
Patch Set 2:

I would say managing the path in a single file is better than spreading the path among modules. An example would be, if we want to refactor the structure of the project(may not happen currently), it's easy to only update the source instead of updating many files, which is less likely to cause regression.

Patch-set: 2
Label: Verified=0