Don't display functest report for single test cases
When executing a single test case, it is not neccessary
to show duplicated information about the result of the test
case. The functest report and the deployment description
only makes sense when running multiple tests. This patch
disables that when running a single test case,
e.g. functest testcse run vping_ssh

The report is still showed when doing:
     functest testcase run all

Change-Id: I92d97d6111ecd9484a14c72cf4bb0c2bbe3ae6bf
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>