Update patch set 4
Patch Set 4:

As the fact that: refstack-client will verify the existence of .tempest which should be in the same dir of refstack-client [1].

Now in functest container, .tempest is generated in /home/opnfv by default, while refstack-client is located in /usr/local/lib/python2.7/dist-packages, so when running refstack_defcore, it will complain that:
 refstack_client:109 ERROR Tempest directory given is not a directory or does not exist: /usr/local/lib/python2.7/dist-packages/.tempest

And this is the only issue which blocks refstack running.

So, there are 2 options for us:
1) Clone refstack-client repository in the Docker image and install .tempest and .venv in the dir of refstack-client;
2) Ask refstack team for help, see if it is possible to update the logic to allow .tempest exist in other dir.
 
So, what's your suggestions? @Jose, @Matthew, @Cedric? 

[1]:https://github.com/openstack/refstack-client/blob/master/refstack_client/refstack_client.py

Patch-set: 4