Update patch set 1
Patch Set 1:

Yes. And there are smaller issues as well (permission rights, bash...).
From now, all the work is useful for both (Ubuntu 14.04/16.04 or Alpine). 

But, at least, we could design an alpine container for all compatible testcases. Badly refstack_core is listed as smoke. But wait and see.

Patch-set: 1