Update patch set 2
Patch Set 2: Code-Review-1

is http still supported in Functest? if so does it mean that even if we are testing http, requests scenario is no longer supported in Functest?
Couldn't it be better that requests' execution determined by http/https, if it is a http condition, requests will be executed, or else it will not, rather than delete it directly?

Patch-set: 2