Update patch set 1
Patch Set 1: Code-Review+1

I tested this locally and it works. 

However, we are installing tempest and at the same time cloning the repo. I have tried to install rally verify from the tempest installation in /usr/local/lib/python2.7/dist-packages/tempest but it doesn't work because it needs a git repo to do so... So, we need the tempest repo for Rally-tempest and refstack to be consumed. I wonder if we still need to install it in requirements...

Patch-set: 1
Label: Code-Review=+1