Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

We should clarify why it's required to clone tempest. At least Functest is not in charge of that as we handle tempest by package for the related TestCase.

I think in case of Rally, it's only required when running tempest which is redundant regarding the related tempest TestCase.

Patch-set: 1
Label: Code-Review=-1