Update patch set 2
Patch Set 2:

this script is using gsutil and artifacts, which has nothing to do with Functest processes. Gsutil should be managed by releng. If this script only applies to functest, then you can rename it to functest_upload_apidocs or similar. 

api_build directory can be a parameter that it could be different for each repo if needed.

Patch-set: 2