New python script to check openstack deployment
This is a crucial part of functest env prepare.
The former check_os.sh is missing some error
control and is sometimes trying to connect to an
empty endpoint if the sed/awk commands were wrong.
This python script is more robust in that sense.
You can see an example of execution in [1].

UT to be added

[1] http://paste.openstack.org/raw/614728/

Change-Id: I622befa13b58d2d31a08d307befb12a2be28fe4d
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>