Update patch set 1
Patch Set 1:

We have noticed in other work that some failed test cases happen e.g. on system deployed with apex installer but pass with fuel installer. Instead of blacklisting a long (in)complete list of scenarios, allowing the use of a wildcard for each scenario field is a cleaner solution.

Patch-set: 1
Label: Verified=0