Update patch set 7
Patch Set 7: Verified-1

> (7 comments)
 > 
 > I don't see the whole point.
 > 
 > If the purpose to get rid of check_os.sh, the change must remove
 > the former file too and everything must be adapted as well
 > (prepare_env, setup.cfg...)
 > If the purpose is also to remove prepare_env, we must configure
 > rally somewhere and adpat jjobs too.
 > 
 > Could you please check it's rated 10/10 by pylint because it's a
 > new trivial script?
 > 
 > BYW, I would propose to vote verify -1 when incomplete patches are
 > published to be shared.

The purpose is to get rid of check_os.sh only, not prepare_env. The change from check_os.sh to this script is done https://gerrit.opnfv.org/gerrit/#/c/37071/ which depends on this one. I want to keep them separated as it's a critical part and too many lines of code to be changes. This patch adds the new code and the other replaces the old usage by the new one.

I did vote verify -1 before but since now jenkins is returning -1 as well, I didn't vote myself this time. Still need to fix a unit test that is failing (help is appreciated).

I will work on the changes you suggest.

Patch-set: 7
Label: Verified=-1