Update patch set 1
Patch Set 1:

I have not been able to test this (it being arm), but putting it up to at least get the ball rolling on how we could do this. Happy to co-author with anyone, as long as we get something in place to verify integrity. Taken from https://github.com/nodejs/node/#verifying-binaries

Patch-set: 1
Label: Verified=0