Update patch set 2
Patch Set 2:

Yes. I think you can't test the change without a trick because it downloads file from master (I apply a hook to download the right ref when building the Alpine images).

Please let me know if it fails when the dependencies are merged. Normally it will be ok.

Patch-set: 2