Update patch set 3
Patch Set 3:

(1 comment)

Jose, Morgan: Let me know if I do manage docker tags. Here this script is designed for master only (it must be modified for stable/xxx as well for Dockerfile and requirements files), then latest is enough. But we could also add master or the gerrit change as well. opnfv (as repo) is forced because it's directly used in Dockerfile dependencies as well.

I could check if pushing is required to avoid hacking for personal testing. But that could be enhanced in next changes.

Patch-set: 3