Add --debug flag to run_tests
This is a proposal to show the log information of a testcase
on the console in real time. It relies on the flag -d/--debug
in run_tests.py
If this flag is provided, it will modify some paramteres in
logging.ini and return to the previous state at the end of
the script execution.

To be done (if proposal is accepted):
- add UT
- add CLI support for the new flag

Output of local testing:
http://paste.openstack.org/raw/615151/

Change-Id: I2f29455b633fcd3d204d6dc38f22178e560b2ab3
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>