Update patch set 6
Patch Set 6:

I will ask them if they can update their image to the version we offer, maybe it's ok for them and we don't need to download another cirros image.

About the checksum, yes. I will do that in another patch and then we fix the security issues with wget.

Patch-set: 6