Update patch set 3
Patch Set 3:

> I don't think docker/Dockerfile.aarch64 hasn't been updated as much
 > as docker/Dockerfile. Could you please check the diff between them?
 > 
 > We are switching to Alpine. I will have a look after my holidays
 > how to take this arch into account in a more suitable way (the
 > current approach of duplicating the Dockerfile is not good and is
 > useless).

Hi Cedric, I agree I don't like this double docker file either. For other projects, we apply a patch in releng with aarch64 diffs over the initial dockerfile (for example https://gerrit.opnfv.org/gerrit/#/c/30185/).

For now we are not impacted by the dockerfile mismatch as we are still using x86 jenkins slaves and therefore the x86 functest image. But we definitely want to be migrate to arm jenkis slaves and have a functest image in sync on both architectures.

Patch-set: 3