Update patch set 19
Patch Set 19:

(1 comment)

Sorry in the last patch we have the problem of mingle api with cli, now we have another, is where to put the abstracted common methods, we cannot put it in cli, it is actually still mingled in this way. In my opinion the best solution should be put them either in the source code or we keep some replica code. Take list test case for example:
1) in the source code define a new method list_summary_tests()(naming is bad just for instance), then call it directly in CliTestcase and ApiTestcase
2) Keep CliTestcase untouched, do it again in ApiTestcase

What do you think Jose?

And my apology for didn't notice this in the first place

Patch-set: 19