Update patch set 3
Patch Set 3:

Yes, maybe improve the tittle is a good idea.
What we are cleaning is not tempest objects, it's the things that we create before running the tests, the network, the flavor, the image and so on.

Patch-set: 3
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0