Update patch set 20
Patch Set 20:

> @Serena, as to exception, if testcase_name is wrong for the url:
 > /api/v1/functest/testcases/<testcase_name>, http 404 exception
 > usually indicates the url is not available, actually it means the
 > test case with this testcase_name does not exist.
 > 
 > It is not reasonable to mix http error with custom error code,
 > cause http exception can not represent all errors in this case
 > correctly.
 > 
 > Why is it wrong to define our own error code?

Hi Linda, the situation you described is not conflicted at all. An URL(Uniform Resource Locator) implies the means to access an indicated resource, if the URL is not accessible(expressed as NOT FOUND), it means the resource is not there, in our case testcase_name required is not exist.

And true, we can not customize the HTTP status code, all are defined in https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html

Patch-set: 20
Label: Verified=0