Update patch set 3
Patch Set 3:

I would go further and ask Doctor team to re-work the run.sh script that they use to trigger the tests... That script is calling python scripts in the end. Why not using directly python with proper calls? Of course, this is just a suggestion from my side, but I agree, at least a python package should be setup in Doctor repo and consumed in Functest (specially to leverage dependencies)

Patch-set: 3
Label: Verified=0