Update patch set 1
Patch Set 1:

I'v already passed the integration test locally, that's why i submitted the patch. Additionally, the latest version of tosca-parser is 0.8.1 and the latest version of heat-translator is 0.9.0 if you install them through pip. Parser will use the latest features of tosca-parser and heat-translator, and parser has implemented some features based on etsi specification, which are not included in tosca-pare or heat-translator.

Patch-set: 1