Update patch set 2
Patch Set 2: Code-Review-1

IMHO, ping once is quite unreliable, because maybe the network link is not ready when the ping starts, in that case, try several times is very necessary. And timeout in another way guarantees that ping will not proceed forever, after several tries, if ping is still not ok, it indicates ping truly fails, in other words, network link will never be established.

Patch-set: 2