Update patch set 1
Patch Set 1:

> Please fix the nit in git commit message (block)
 > 
 > I do think they are fundamental in case of ODL scenarii. My main
 > concern is related to the bug regarding keystone v3 instead of the
 > blocking status.

I think Nikolas would like to move to non bloking to see the results of test cases from feature tests that cannot be currently run due to the fact that an ODL scenario is required but as the odl test if FAIL (due to keystone v3) and blocking, all odl subsequent tests are not run due to the bloking status. So the test of the patch is maybe misleading, it is not fundamental regarding subsequent tests.

Patch-set: 1
Label: Code-Review=0