Update patch set 2
Patch Set 2:

> (1 comment)

Hi Tim, 
Sorry for the second comment. 
The IP address 192.168.0.2 is the admin network IP for the undercloud. We are not sure what is the default IP used in the CI environment. If you could provide us that IP we can change it in our scripts as well. 
Thanks!

Patch-set: 2