Add UEFI support in functest
Add extra properties to the cirros image
for ARM architecture.
In the glanceclient repository, the create method
expects **kwargs as parameters, thus an extra_properties
variable containing a dictionary of property-value pairs
can be added (if given) to the image to be created.
If the properties do not appear, the call should not affect
the image properties because the dictionary is initialized
as empty.

Change-Id: I690da90e6c6f250bd0a3d99cce39aa250e19b9c2
Signed-off-by: Catalina Focsa <catalina.focsa@enea.com>