Update patch set 1
Patch Set 1:

The request to our IT department is still open to change the marketplace port to 8080. I think that's the problem since 8082 is blocked on most pods, right? We'll try to fix asap. The image is actually used by openims testcase only but the descriptors are downloaded as well from the marketplace. So both testcases are effected in the end.

Patch-set: 1
Reviewer: Gerrit User 6545 <6545@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0