Update patch set 4
Patch Set 4:

Yes, you're right: the shebang is correct when using console script.
My main concern is if we create a separate docker container for doctor (python3) as we are doing for parser. 
Doctor-tests is not fully conformed with OPNFV E requirements and it could raise side effects.

Patch-set: 4
Label: Verified=0