Enable Domino testcase
Domino is part of the OPNFV projects released for E [1].
As now Domino is ready for Functest integration (package and
requirements), the testcase can be enabled.

If the testcase fails, the issues should be related to Domino.

[1] https://lists.opnfv.org/pipermail/opnfv-tech-discuss/2017-September/017951.html

Change-Id: I26f3d3ae338def0f77c46c2cbda1baf3d55c1a41
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>
(cherry picked from commit 555ac3f8656004d469c04370448a8a46f8a1418d)