Update patch set 8
Patch Set 8:

> Yes it will be great to create the jenkins jobs instead of using
 > Docker automated builds which were created only for my own
 > repository (testing purpose).
 > 
 > But I am not convinced that storeperf's proposal is the best one.
 > First it forbids the classical builds (FROM in third line + ARG in
 > image) and I would have thought using MANIFEST instead. But I
 > couldn't take time to go deeply into details.

Hi,
I am not that familiar with Docker, so I'll let Delia investigate the possibility of using MANIFEST.
As for the job definitions (multijob + phases for each container derived from functest-core), we (Armband) can put something together in a week or two, if you're ok with it. It would help us create AArch64 jobs in parallel to x86_64 ones, since the current approach is trickier to adapt.

The way this looks (and the issues we saw in building the Alpine containers manually on AArch64) makes me think we will stick to Ubuntu Docker for the 5.0.0 release. We do want to get on the Alpine train asap, but we don't want to sidetrack you from actual Functest work.
BR, Alex

Patch-set: 8
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0