Update patch set 1
Patch Set 1:

agree with previous comments

if the argument is to save CI processing (which is a good argument as vnf tier is longer, it used to be in weekly loop that never happened in previous version), it shall be managed at releng level not at functest level

please note that the vnf tier is only invoked on 1 scenario os-nosdn-nofeature-ha, which is supposed to be the more stable and up&running since MS3.
The choice to run it on only 1 tier was also to save CI resources.

It seems that the cases could work on fuel on x86, there is a connectivity issue between the POD and the server hosting some images for such test cases

Patch-set: 1
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0