Update patch set 3
Patch Set 3:

shall we add also a section on how to use it?
you need to edit the custom config file to precise which cases you want to run, this could be documented, no?
and then even when it was referenced in the Default testcases.yaml, you had to precise that before running the test, it means you have to do manual config anyway

another possible option would be to create testcases-debug.yaml under the different docker with 
- tempest_custom (in smoke)
- no constraint on all the possible cases

through --debug option it should even be possible to use this yaml rather that the default one
and we could do the same with a logging-debug.ini

Patch-set: 3
Label: Code-Review=0