Update patch set 1
Patch Set 1:

The testcase implementation is fully defined in functest. But I agree that you're in charge of the robot files.

I'm quite ok to switch the project name but the git commit message could be more precised. If the database has already been modified, we should accept the change.

Patch-set: 1
Label: Code-Review=0