Update patch set 1
Patch Set 1: Code-Review+1

if INSTALLER_TYPE is None, it means non OPNFV solution
it would be up to the tester to adapt the contraints according to his/her knowledge of his/her solution

we defined our constraints according to our scenarios/installers
E.g. For a test case, if a constraint is apex, we may assume that it may be compatible with redhat based installer, so it may make sense to keep it running. Only the tester who set None to the env var know

Patch-set: 1