Update patch set 10
Patch Set 10:

I think it may be safely abandonned. Only functest-core Dockerfile needs to be modified between amd64 and arm64 in case of cross-building. Else we can simply rely on manifests.

It should be noted that CI is not ready to produce the manifests. As the ARM container is built on top of functest (the former container) we cannot consider that as an issue.

Patch-set: 10