Update patch set 1
Patch Set 1:

Yes, all migration tests (both cold and live) needs to be blacklisted if there is only one compute. Currently the functionality is named as 'no_live_migration' in blacklist (earlier, this used to be set if the live migration feature was not enabled in tempest config file).

This could be now renamed as 'no_migration'; in general, when there are at least two computes, both cold and live migration are supported.

Another thing is that at the moment rally_sanity/full live migration tests will fail if the hosts are not using shared storage (relates to https://bugs.launchpad.net/rally/+bug/1721214).

Patch-set: 1
Label: Code-Review=0