Update patch set 2
Patch Set 2:

Yes. It can be merged in master for gating but we should wait several runs before the cherry-pick. Technically speaking, only details are lost which is conformed with BashFeature. Then it's not a bugfix and we can safely wait a little bit.

I think you could apply the same rules in sdnvpn.

Patch-set: 2