Update patch set 1
Patch Set 1: Code-Review-1

Is it required by Functest or by a third-party project?
It would be better to implement this methods in the right place.

It should be noted that we could also get rid of that utils.

Patch-set: 1
Reviewer: Gerrit User 3405 <3405@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1