Update patch set 1
Patch Set 1:

Cedric is right. openstack_utils was introduced in Brahmaputra to make easier the creation of openstack objects. However, we didn't have proper control and allowed everyone to add code in that file. The result is that in the end each function is completely different from the others when it comes to return values, error handling, etc.. Besides, the coverage is poor... 
Snaps is more mature in this aspect and you are free to use it in sdnvpn, or just the python client as Cedric says (for this case it is simpler and easier).

As a recommendation, I wouldn

Patch-set: 1