Update Functest to OpenStack stable/pike
It updates setup.py, (test-)requirement.txt and upper-constraints.txt
according to OpenStack requirements [1].

It disables all Features testcases as they are not synchronized with
OpenStack stable/pike yet [2].

The docker functest-parser is simplified as it can be based on
functest-core as both Functest and Parser leverage on OpenStack
stable/pike.

ODL_TAG and REFSTACK_TAG variables are set to master again.

[1] http://git.openstack.org/cgit/openstack/requirements/log/?h=stable/pike
[2] https://wiki.opnfv.org/display/functest/Requirements+management

Change-Id: I23fa84f35244c1f4026abd89414b99576f042ee4
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>