Update patch set 2
Patch Set 2:

Yes I'm fine with moving the operations for ARM PODs in run_test as proposed. It allows modifying the jenkins jobs for the new target (run_tests as a single entry point for running tests).

It will be great if you can write the UT as proposed here and adapt the jenkins jobs as well 
[1] https://gerrit.opnfv.org/gerrit/#/c/47259/

Maybe we will enhance run_tests.py a little bit in a second step by moving few operations to the TestCase class.

Patch-set: 2
Label: Verified=0