Update patch set 1
Patch Set 1: Code-Review-1

In my humble opinion, large page is not required for all pods or environments, and the default configuration is not leveraging huge page.
And the test cases should be suitable for as many situations as possible, in case of the development env or virtual pod, which the resource is limited, these test cases should also work fine. 
My suggestion is instead of simply changing the value, which maybe will induce other problems in not-compass-huge-page env, maybe the better way to solve it is by configuring it according to different scenarios(huge page or normal condition).

Patch-set: 1
Reviewer: Gerrit User 2651 <2651@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1