Update patch set 1
Patch Set 1:

I would have prefered to leverage on tempest logic.
Could you please confirm that tempest is not able to detect the target and only execute the right tests? Or simply if it considers by default the double version support.

If it only concerns Rally, this logic could be in the Rally testcases.

Else I'm fine with this proposal.

Patch-set: 1