Build vnf and restapi containers for ARM
The root cause of the problem was the installation
of unf_ext packet. The Gemfile.lock in
https://github.com/boucherv-orange/clearwater-live-test
asked for version 0.0.6 and this version was not found for arm.
When it tried to build the package, the conversion from int
to char error was thrown (issue reported here
https://github.com/knu/ruby-unf_ext/issues/15).

As now version 0.0.7.4 is selected, then the vnf docker
images are built without any other errors.

Gemfile.lock was updated in the original repo,
this fix enables vnf and restapi containers build for ARM.

JIRA: FUNCTEST-899

Change-Id: Id2c5fe82cb3705f58a8175fcf0a17f443e9b7d9d
Signed-off-by: Delia Popescu <delia.popescu@enea.com>