Fix vnf and restapi Dockerfile for ARM
The root cause of the problem is the installation
of unf_ext packet. The Gemfile.lock in
https://github.com/boucherv-orange/clearwater-live-test
asks for version 0.0.6 and this version is not found for arm.
When it tries to build the package, the conversion from int
to char error is thrown (issue reported here
https://github.com/knu/ruby-unf_ext/issues/15).
If version 0.0.7.4 is used, then the vnf docker
image is built without any other errors.
JIRA: FUNCTEST-899

Change-Id: Id2c5fe82cb3705f58a8175fcf0a17f443e9b7d9d
Signed-off-by: Delia Popescu <delia.popescu@enea.com>