Update patch set 4
Patch Set 4:

Yes it should be functest-core which is based on alpine + functest python package and its requirements. 
here we could create a virtualenv if required but the process is not the best (pip is useless here) and we could leverage on system librairies as all dependencies are met.

Patch-set: 4