Update patch set 3
Patch Set 3:

We're creating a testing network to plug testing vms to it, no?
I fully agree there is a bug here, but I tried to highlight that this patchset is needed only if we couldn't set the right network to the testing vms.

I know very well what shared means in the OpenStack context.

Patch-set: 3