Update patch set 1
Patch Set 1:

I agree with your point but it would ease checking the disabled rules at the end.
Here we could also set docstrings instead of the method name if we want to be very precised.

I'm fine to merge it asis. This patch is already an added value.

Patch-set: 1