Update patch set 1
Patch Set 1:

> giga (wget opts) or mega (commit message) ? The best here is per
 > 100M :)

it's 1M per dot, leading to 1G per ? lines, which is called 'dot:giga' in wget. Not very intuitive ...
We can do any combination, see [1].

 > I would prefer progess bar here which works well if TTY.
 > 
 > Is the issue related to a jenkins job downloading the files?
 > Normally they should be downloaded out of jobs.

This is not a real issue (nothing is broken), just that it prints a lot of dots when downloading big files, which I find annoying.
I'm ok with progress bar or any presentation, as long as it doesn't output 10K lines of log while downloading a single file.

I'm not that interested in merging this, so if you think it's too much trouble, we can leave it be.
I just did a similar change in Fuel recently and thought I should share my 'dot:giga' findings :)

[1] https://serverfault.com/a/225397

Patch-set: 1